Fixed AppConstants to store a single Guid instance
